✨ (Background Context Menu) - Add Background Color Context Menu
