+++ src/app/image-generator/_components/section/background-color/options/TransparentButton.tsx
@@ -1,38 +1,13 @@
 import { Button } from "@/components/ui/button";
-import { useImageGeneratorStore } from "@/lib/store/imageGenerator.store";
+import { setTransparentBackground } from "@/lib/utils/image-generator/setBackgroundColor/setTransparentBackground";
 import { Eraser } from "lucide-react";
 
 const TransparentButton = () => {
-  const gradient = useImageGeneratorStore(
-    (s) => s.settings.background.gradient
-  );
-
-  const setBackground = useImageGeneratorStore(
-    (s) => s.setBackground
-  );
-  const setDimensions = useImageGeneratorStore((s) => s.setDimensions);
-
-  const resetBackground = useImageGeneratorStore((s) => s.resetBackground);
-
-  const handleClick = () => {
-    resetBackground();
-    setDimensions({ format: "png" });
-    setBackground({
-      backgroundMode: "solid",
-      backgroundColor: "",
-      tailwindColor: "",
-      gradient: {
-        ...gradient,
-        useVia: false,
-      },
-    })
-  };
-
   return (
     <Button
       variant="outline"
       size="lg"
-      onClick={() => handleClick()}
+      onClick={() => setTransparentBackground()}
       className="mx-1 w-full"
     >
       <Eraser className="mr-2 size-4" />
