+++ src/lib/utils/image-generator/setBackgroundColor/setRandomCustomBackground.ts
@@ -1,0 +1,17 @@
+import { defaultImageGeneratorSettings } from "@/lib/constant/defaultImageGeneratorSettings";
+import { useImageGeneratorStore } from "@/lib/store/imageGenerator.store";
+import { getRandomColor } from "../../colors/getRandomColor";
+
+export const setRandomCustomBackground = () => {
+  const setBackground = useImageGeneratorStore.getState().setBackground;
+  const magicColor =
+    useImageGeneratorStore.getState().settings.background.magicColor;
+  const defaultBackgroundSettings = defaultImageGeneratorSettings.background;
+
+  const randomColor = getRandomColor();
+  setBackground({
+    ...defaultBackgroundSettings,
+    backgroundColor: randomColor.hex,
+    magicColor
+  });
+};
